refact: moved spot card to own component
